[#153] Add hooks utils to contain postProcess
