+++ .tool-versions
@@ -1,1 +1,2 @@
 nodejs 18.12.1
+terraform 1.3.9
