+++ src/utils/hooks.test.ts
@@ -1,0 +1,55 @@
+import { remove } from '@/helpers/file';
+import { formatCode, detectTerraform } from '@/helpers/terraform';
+
+import { postProcess } from './hooks';
+
+jest.mock('@/helpers/terraform');
+
+describe('postProcess', () => {
+  const projectDir = 'postProcess-test';
+  const generalOptions = { projectName: projectDir, provider: 'aws' };
+
+  describe('given current machine had terraform', () => {
+    beforeAll(async () => {
+      (detectTerraform as jest.Mock).mockImplementation(() => true);
+
+      await postProcess(generalOptions);
+    });
+
+    afterAll(() => {
+      jest.resetAllMocks();
+      remove('/', projectDir);
+    });
+
+    it('runs formatCode', async () => {
+      await expect(formatCode).toHaveBeenCalled();
+    });
+  });
+
+  describe('given current machine did not have terraform', () => {
+    const consoleErrorSpy = jest.spyOn(global.console, 'error');
+
+    beforeAll(async () => {
+      (detectTerraform as jest.Mock).mockImplementation(() => {
+        throw new Error('terraform not found');
+      });
+
+      await postProcess(generalOptions);
+    });
+
+    afterAll(() => {
+      jest.resetAllMocks();
+      remove('/', projectDir);
+    });
+
+    it('does NOT run formatCode', async () => {
+      await expect(formatCode).not.toHaveBeenCalled();
+    });
+
+    it('displays the error message', () => {
+      expect(consoleErrorSpy).toHaveBeenCalledWith(
+        Error('terraform not found')
+      );
+    });
+  });
+});
